% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkMaxSpan}
\alias{checkMaxSpan}
\title{checkMaxSpan
Check the maxSpan element, for Time objects}
\usage{
checkMaxSpan(maxSpan)
}
\arguments{
\item{maxSpan}{The list of ranges}
}
\description{
checkMaxSpan
Check the maxSpan element, for Time objects
}
