% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkRanges}
\alias{checkRanges}
\title{checkRanges
Check the ranges element, for Date objects}
\usage{
checkRanges(ranges)
}
\arguments{
\item{ranges}{The list of ranges}
}
\description{
checkRanges
Check the ranges element, for Date objects
}
