% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterEMPTY}
\alias{filterEMPTY}
\title{filterEMPTY
Filter empty elements of a list}
\usage{
filterEMPTY(x)
}
\arguments{
\item{x}{The list of values}
}
\description{
filterEMPTY
Filter empty elements of a list
}
