% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeInput}
\alias{makeInput}
\title{makeInput
Make the input div-tag}
\usage{
makeInput(label, inputId, class, icon, style, options)
}
\arguments{
\item{label}{The label of the daterangepicker}

\item{inputId}{The inputId of the daterangepicker}

\item{class}{The class of the daterangepicker}

\item{icon}{The icon of the daterangepicker}

\item{style}{The style of the daterangepicker}

\item{options}{The options of the daterangepicker}
}
\description{
makeInput
Make the input div-tag
}
