% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeLabel}
\alias{makeLabel}
\title{makeLabel
Make the label}
\usage{
makeLabel(label, inputId)
}
\arguments{
\item{label}{The label of the daterangepicker}

\item{inputId}{The inputId of the daterangepicker}
}
\description{
makeLabel
Make the label
}
