% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daterangepicker.R
\name{updateDaterangepicker}
\alias{updateDaterangepicker}
\title{updateDaterangepicker
Change the start and end values of a daterangepicker on the client}
\usage{
updateDaterangepicker(
  session,
  inputId,
  label = NULL,
  start = NULL,
  end = NULL,
  minYear = NULL,
  maxYear = NULL,
  icon = NULL
)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{start}{The start date. Either a Date object, or a string in yyyy-mm-dd format.}

\item{end}{The end date. Either a Date object, or a string in yyyy-mm-dd format.}

\item{minDate}{The minimum allowed date. Either a Date object, or a string in yyyy-mm-dd format.}

\item{maxDate}{The maximum allowed date. Either a Date object, or a string in yyyy-mm-dd format.}
}
\description{
updateDaterangepicker
Change the start and end values of a daterangepicker on the client
}
