% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daterangepicker.R
\name{daterangepicker}
\alias{daterangepicker}
\title{daterangepicker}
\usage{
daterangepicker(
  inputId = NULL,
  label = "Select a Date",
  start = NULL,
  end = NULL,
  min = NULL,
  max = NULL,
  ranges = NULL,
  language = "en",
  style = "width:100\%;border-radius:4px;text-align:center;",
  class = NULL,
  icon = NULL,
  options = daterangepickerOptions(),
  initCallback = NULL
)
}
\arguments{
\item{inputId}{(string) The input ID}

\item{label}{(string) The label for the control, or NULL for no label.}

\item{start}{(Date or string) The beginning date of the initially selected
date range. If you provide a string, it must match the date format string
set in your locale setting.}

\item{end}{(Date or string) The end date of the initially selected date
range.}

\item{min}{(Date or string) The earliest date a user may select.}

\item{max}{(Date or string) The latest date a user may select.}

\item{ranges}{(List) Set predefined date ranges the user can select from.
Each key is the label for the range, and its value an array with two dates
representing the bounds of the range. Click ranges in the configuration
generator for examples.}

\item{language}{(string) The language used for month and day names. Default
is "en". See the \href{https://momentjs.com/}{Multipe Locale Support} for a
list of other valid values.}

\item{style}{(string) Add CSS-styles to the input.}

\item{class}{(string) Custom class}

\item{icon}{(shiny.tag) Icon to display next to the label.}

\item{options}{(List) List of further options. See
\code{\link{daterangepickerOptions}}}

\item{initCallback}{(function) A JS function}
}
\description{
Date Range Picker can be attached to any webpage element to pop up two
calendars for selecting dates, times, or predefined ranges like "Last 30
Days"
}
\seealso{
\href{https://www.daterangepicker.com/#config}{www.daterangepicker.com}
}
