% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daterangepicker.R
\name{daterangepicker}
\alias{daterangepicker}
\title{daterangepicker}
\usage{
daterangepicker(
  inputId = NULL,
  label = "Select a Date",
  start = NULL,
  end = NULL,
  min = NULL,
  max = NULL,
  ranges = NULL,
  language = "en",
  style = "width:100\%;border-radius:4px;text-align:center;",
  class = NULL,
  icon = NULL,
  options = daterangepickerOptions(),
  initCallback = NULL
)
}
\arguments{
\item{inputId}{The input ID}

\item{label}{The label for the control, or NULL for no label.}

\item{start}{The beginning date of the initially selected. Must be a Date or
string.}

\item{end}{The end date of the initially selected date range. Must be a Date
or string}

\item{min}{The earliest date a user may select. Must be a Date or string}

\item{max}{The latest date a user may select. Must be a Date or string}

\item{ranges}{Set predefined date ranges the user can select from. Each key
is the label for the range, and its value an array with two dates
representing the bounds of the range.}

\item{language}{The language used for month and day names. Default is "en".
See the \href{https://momentjs.com/}{Multipe Locale Support} for a list of
other valid values.}

\item{style}{Add CSS-styles to the input.}

\item{class}{Custom class}

\item{icon}{Icon to display next to the label.}

\item{options}{List of further options. See
\code{\link{daterangepickerOptions}}}

\item{initCallback}{A \code{\link[htmlwidgets]{JS}} function}
}
\description{
The Date Range Picker pops up two calendars for selecting dates, times, or
predefined ranges like "Yesterday", "Last 30 Days", etc.
}
\examples{
if (interactive()) {
library(shiny)
library(daterangepicker)

## UI ##########################
ui <- fluidPage(
  tags$head(tags$style(".myclass {background-color: #96dafb;}")),
  daterangepicker(
    inputId = "daterange",
    label = "Pick a Date",
    start = Sys.Date() - 30, end = Sys.Date(),
    max = Sys.Date(),
    language = "en",
    style = "width:100\%; border-radius:4px",
    class = "myclass",
    icon = icon("calendar")
  ),
  verbatimTextOutput("print"),
  actionButton("act", "Update Daterangepicker"),
)

## SERVER ##########################
server <- function(input, output, session) {
  output$print <- renderPrint({
    req(input$daterange)
    input$daterange
  })
  observeEvent(input$act, {
    updateDaterangepicker(session, "daterange",
                          start = Sys.Date(),
                          end = Sys.Date() - 100)
  })
}
shinyApp(ui, server)
}
}
\seealso{
\href{https://www.daterangepicker.com/#config}{www.daterangepicker.com}

Other daterangepicker: 
\code{\link{daterangepickerOptions}()},
\code{\link{updateDaterangepicker}()}
}
\concept{daterangepicker}
