% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daterangepicker.R
\name{updateDaterangepicker}
\alias{updateDaterangepicker}
\title{updateDaterangepicker}
\usage{
updateDaterangepicker(
  session,
  inputId,
  label = NULL,
  start = NULL,
  end = NULL,
  min = NULL,
  max = NULL,
  icon = NULL,
  options = NULL
)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{start}{The start date. Either a Date object, or a string in yyyy-mm-dd
format.}

\item{end}{The end date. Either a Date object, or a string in yyyy-mm-dd
format.}

\item{min}{(Date or string) The earliest date a user may select.}

\item{max}{(Date or string) The latest date a user may select.}

\item{icon}{(shiny.tag) Icon to display next to the label.}

\item{options}{(List) List of further options. See
\code{\link{daterangepickerOptions}}}
}
\description{
Change the start and end values of a daterangepicker on the client
}
